backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
black==24.3.0
django-ses==4.1.0
boto3==1.35.0
django-storages==1.14.4